Endrer flere filendinger til .jsx
