+++ src/components/SlateEditor/plugins/table/schema.jsx
