+++ src/containers/StructurePage/__tests__/AddResourceModal-test.jsx
