+++ src/containers/StructurePage/__tests__/ConnectFilters-test.jsx
