+++ src/containers/StructurePage/__tests__/EditFilters-test.jsx
