+++ src/containers/StructurePage/__tests__/InlineAddButton-test.jsx
