+++ src/containers/StructurePage/__tests__/InlineEditField-test.jsx
