+++ src/containers/StructurePage/__tests__/ResourceItems-test.jsx
