+++ src/containers/StructurePage/__tests__/StructurePage-test.jsx
